+++ apps/cowswap-frontend/src/modules/accountProxy/utils/sumUpUsdAmounts.ts
